requirements.txt
APScheduler==3.9.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
idna==3.3
numpy==1.23.2
pandas==1.4.4
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pythonw==3.0.3
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
six==1.16.0
slack-bolt==1.14.3
slack-sdk==3.18.3
tabulate==0.8.10
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
